added API key to the files in folder Views and moved API file to views
