+++ news/src/Views/Weather.jsx
@@ -1,15 +1,27 @@
 import React from 'react';
 import Article from '../Components/Article';
+import API from "./API";
 
 function Weather() {
+    const [articles, setArticles] = useState([]);
+  
+    useEffect(() => {
+      API.getWeather().then((results) => {
+        setArticles(selectTopFive(results.data.articles));
+      });
+    }, []);
+  
     return (
-        <div>
-            <h1>Weather News goes here: </h1>
-            <Article/>
-            <Article/>
-            <Article/>
-        </div>
-    )
-}
+      <>
+        <h1>Weather: </h1>
+        {articles.map((article, i) => (
+          <Article
+          key={i}
+          data={article[i]}
+        />
+        ))}
+      </>
+    );
+  }
 
 export default Weather;
